Door Mechanics, Bathroom/TV Furniture
Added Scripts that open doors when 'Activated' - that's how we'll open
doors remotely in all rooms. For now, they just open when the player
walks up to them. Added Furniture blocks to bathroom and TV room, but
still need to put on texture meshes. I'll keep working on this over the
next few days, so let me know if you're gonna work on something so I can
commit and then pull your changes when you're done.